requirements.txt
# --- Runtime ---
numpy>=1.24.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tenacity>=8.2.0

# --- Agent framework ---
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.7.0

# --- PDF parsing ---
pymupdf>=1.24.0

# --- API layer ---
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.1.0
python-multipart>=0.0.9

# --- Data collection only (install separately: pip install -r requirements-collection.txt) ---
# requests, beautifulsoup4, lxml, tqdm, sentence-transformers
